Fixed: Octal print errors

+++ printer.c
@@ -61,10 +61,10 @@ static int raise(int n)
 */
 int __write_number(va_list args, int *printed)
 {
-	int value, temp, digits;
+	long int value, temp, digits;
 
 
-	value = va_arg(args, int);
+	value = va_arg(args, long int);
 	digits = 0;
 
 	if (value < 0) /* If this number is negative */
